// ng-client/src/app/interfaces/posts.interface.ts
import { ICategories } from "./categories.interface";
import { IComments } from "./comments.interface";


export interface IPosts {
    id?: number;
    author?: string;
    content?: string;
    thumbnail?: string;
    createdAt?: string;
    title?: string;
    picture_author_post?: string;
    comments?: IComments[];
    likes?: number;
    dislikes?: number;
    categories?: ICategories[];
}